requirements.txt
africastalking==1.2.5
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
altair==5.2.0
amqp==5.1.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Authlib==1.2.1
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
blinker==1.7.0
boto3==1.21.8
botocore==1.24.46
cachetools==5.3.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contextlib2==21.6.0
contourpy==1.2.0
cryptography==41.0.3
cycler==0.12.1
dataclasses-json==0.5.14
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
dnspython==2.4.2
docarray==0.32.1
elevenlabs==0.2.8
emoji==2.8.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==1.2.0
Faker==21.0.0
favicon==0.7.0
filetype==1.2.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
flask-setup==0.5.5
Flask-SQLAlchemy==3.0.3
flower==2.0.1
fonttools==4.47.0
frozendict==2.3.10
frozenlist==1.3.3
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==20.1.0
h11==0.14.0
hnswlib==0.8.0
htbuilder==0.6.2
httpcore==1.0.2
httpx==0.26.0
humanize==4.8.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==6.1.1
ipython==8.13.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
kombu==5.2.4
langchain==0.1.1
langchain-community==0.0.15
langchain-core==0.1.15
langchain-google-genai==0.0.5
langchain-openai==0.0.3
langsmith==0.0.83
llama-index==0.9.19
loguru==0.7.2
lxml==4.9.3
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
more-itertools==10.1.0
multidict==6.0.4
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numexpr==2.8.6
numpy==1.25.2
openai==0.27.2
orjson==3.9.10
packaging==23.2
pandas==2.1.4
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pexpect==4.8.0
pgvector==0.2.2
pickleshare==0.7.5
Pillow==10.0.1
pinecone-client==3.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.1
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.7
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.4.0
pymdown-extensions==10.5
pyparsing==3.1.1
pypdf==3.16.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2022.7.1
PyYAML==6.0.1
redis==4.5.4
referencing==0.32.0
regex==2023.8.8
requests==2.28.2
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
s3transfer==0.5.2
schema==0.7.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.7
st-annotated-text==4.0.1
stack-data==0.6.2
streamlit==1.29.0
streamlit-aggrid==0.3.4.post3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-javascript==0.1.5
streamlit-keyup==0.2.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.0.5
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
trulens-eval==0.19.2
typer==0.7.0
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.10.15
urllib3==1.26.15
validators==0.22.0
vine==5.0.0
watchdog==3.0.0
wcwidth==0.2.6
websockets==12.0
Werkzeug==2.2.3
wrapt==1.16.0
yarl==1.8.2
zipp==3.15.0
